% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.data.R
\docType{data}
\name{example.data}
\alias{example.data}
\title{example.data}
\format{
A (100x2) 'data.frame' object
\describe{
\item{id}{a data-series}
\item{ans}{another data-series}
}
}
\source{
I made this...
}
\usage{
example.data
}
\description{
splat_packed data
}
\keyword{datasets}
