% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lurch-package.R
\name{lurch-package}
\alias{lurch-package}
\alias{lurch}
\alias{"_PACKAGE"}
\title{lurch: A bare-bones R package}
\description{
A bare-bones R Package build used to test code, functions, data,
imports, etc... Can be a bit of a messy build at times...
}
\details{
This package was splat packed.
}
\section{package functions}{

\link{to be added}
}

